Time for a rewrite, lol.
Took too long of a break + I want to change the approach to the current
project as the goal is not to make a """"game"""" but to test and
compare approaches.

At the top of the list of things to do rn is adding ImGui because this
is absolute cancer to work with.

This commit is so that i can reference things later on if I get stuck.